move `impl pallet_treasury::Config for Runtime` to separate file
